+++ Minio/ApiEndpoints/BucketOperations.cs
@@ -46,9 +46,10 @@ public partial class MinioClient : IBucketOperations
     public async Task<ListAllMyBucketsResult> ListBucketsAsync(
         CancellationToken cancellationToken = default)
     {
-        var requestMessageBuilder = await CreateRequest(HttpMethod.Get).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(HttpMethod.Get).ConfigureAwait(false);
         using var response =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
 
         var bucketList = new ListAllMyBucketsResult();
@@ -72,9 +73,10 @@ public partial class MinioClient : IBucketOperations
         args?.Validate();
         try
         {
-            var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+            var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
             using var response =
-                await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+                await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                        cancellationToken: cancellationToken)
                     .ConfigureAwait(false);
         }
         catch (InternalClientException ice)
@@ -104,8 +106,8 @@ public partial class MinioClient : IBucketOperations
     public async Task RemoveBucketAsync(RemoveBucketArgs args, CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
-        using var response = await this.ExecuteTaskAsync(NoErrorHandlers,
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
+        using var response = await this.ExecuteTaskAsync(ResponseErrorHandlers,
             requestMessageBuilder, cancellationToken: cancellationToken).ConfigureAwait(false);
     }
 
@@ -129,9 +131,10 @@ public partial class MinioClient : IBucketOperations
             args.Location = Config.Region;
 
         args.IsBucketCreationRequest = true;
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var response =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
     }
 
@@ -151,9 +154,10 @@ public partial class MinioClient : IBucketOperations
     {
         args?.Validate();
 
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var responseResult =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
 
         var versioningResponse = new GetVersioningResponse(responseResult.StatusCode, responseResult.Content);
@@ -174,8 +178,8 @@ public partial class MinioClient : IBucketOperations
     public async Task SetVersioningAsync(SetVersioningArgs args, CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
-        using var response = await this.ExecuteTaskAsync(NoErrorHandlers,
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
+        using var response = await this.ExecuteTaskAsync(ResponseErrorHandlers,
             requestMessageBuilder, cancellationToken: cancellationToken).ConfigureAwait(false);
     }
 
@@ -270,9 +274,10 @@ public partial class MinioClient : IBucketOperations
         if (args is null)
             throw new ArgumentNullException(nameof(args));
 
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var responseResult =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
         var getBucketNotificationsResponse =
             new GetBucketNotificationsResponse(responseResult.StatusCode, responseResult.Content);
@@ -298,9 +303,10 @@ public partial class MinioClient : IBucketOperations
         if (args is null)
             throw new ArgumentNullException(nameof(args));
 
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var response =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
     }
 
@@ -320,9 +326,10 @@ public partial class MinioClient : IBucketOperations
         if (args is null)
             throw new ArgumentNullException(nameof(args));
 
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var response =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
     }
 
@@ -352,11 +359,12 @@ public partial class MinioClient : IBucketOperations
             {
                 using var cts = CancellationTokenSource.CreateLinkedTokenSource(cancellationToken, ct);
                 var requestMessageBuilder =
-                    await CreateRequest(args).ConfigureAwait(false);
+                    await this.CreateRequest(args).ConfigureAwait(false);
                 args = args.WithNotificationObserver(obs)
-                    .WithEnableTrace(trace);
+                    .WithEnableTrace(Config.TraceHttp);
                 using var response =
-                    await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+                    await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                            cancellationToken: cancellationToken)
                         .ConfigureAwait(false);
                 cts.Token.ThrowIfCancellationRequested();
             });
@@ -375,9 +383,10 @@ public partial class MinioClient : IBucketOperations
     public async Task<Tagging> GetBucketTagsAsync(GetBucketTagsArgs args, CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var responseResult =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
         var getBucketNotificationsResponse =
             new GetBucketTagsResponse(responseResult.StatusCode, responseResult.Content);
@@ -399,9 +408,10 @@ public partial class MinioClient : IBucketOperations
         CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var restResponse =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
     }
 
@@ -419,9 +429,10 @@ public partial class MinioClient : IBucketOperations
         CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var responseResult =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
         var getBucketEncryptionResponse =
             new GetBucketEncryptionResponse(responseResult.StatusCode, responseResult.Content);
@@ -443,9 +454,10 @@ public partial class MinioClient : IBucketOperations
         CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var restResponse =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
     }
 
@@ -463,9 +475,10 @@ public partial class MinioClient : IBucketOperations
     public async Task SetBucketTagsAsync(SetBucketTagsArgs args, CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var restResponse =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
     }
 
@@ -483,9 +496,10 @@ public partial class MinioClient : IBucketOperations
     public async Task RemoveBucketTagsAsync(RemoveBucketTagsArgs args, CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var restResponse =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
     }
 
@@ -508,9 +522,10 @@ public partial class MinioClient : IBucketOperations
         CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var restResponse =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
     }
 
@@ -529,9 +544,10 @@ public partial class MinioClient : IBucketOperations
         CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var responseResult =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
         var resp = new GetObjectLockConfigurationResponse(responseResult.StatusCode, responseResult.Content);
         return resp.LockConfiguration;
@@ -553,9 +569,10 @@ public partial class MinioClient : IBucketOperations
         CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var restResponse =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
     }
 
@@ -577,9 +594,10 @@ public partial class MinioClient : IBucketOperations
         CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var restResponse =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
     }
 
@@ -597,9 +615,10 @@ public partial class MinioClient : IBucketOperations
         CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var responseResult =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
         var response = new GetBucketLifecycleResponse(responseResult.StatusCode, responseResult.Content);
         return response.BucketLifecycle;
@@ -620,9 +639,10 @@ public partial class MinioClient : IBucketOperations
         CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var restResponse =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
     }
 
@@ -641,9 +661,10 @@ public partial class MinioClient : IBucketOperations
         CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var responseResult =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
         var response = new GetBucketReplicationResponse(responseResult.StatusCode, responseResult.Content);
         return response.Config;
@@ -667,9 +688,10 @@ public partial class MinioClient : IBucketOperations
         CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var restResponse =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
     }
 
@@ -688,9 +710,10 @@ public partial class MinioClient : IBucketOperations
         CancellationToken cancellationToken = default)
     {
         args?.Validate();
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var restResponse =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
     }
 
@@ -733,9 +756,10 @@ public partial class MinioClient : IBucketOperations
         if (args is null)
             throw new ArgumentNullException(nameof(args));
 
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var responseResult =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
         var getPolicyResponse = new GetPolicyResponse(responseResult.StatusCode, responseResult.Content);
         return getPolicyResponse.PolicyJsonString;
@@ -755,9 +779,10 @@ public partial class MinioClient : IBucketOperations
         if (args is null)
             throw new ArgumentNullException(nameof(args));
 
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var response =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
     }
 
@@ -775,9 +800,10 @@ public partial class MinioClient : IBucketOperations
         if (args is null)
             throw new ArgumentNullException(nameof(args));
 
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var response =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
     }
 
@@ -793,9 +819,10 @@ public partial class MinioClient : IBucketOperations
     private async Task<Tuple<ListBucketResult, List<Item>>> GetObjectListAsync(GetObjectListArgs args,
         CancellationToken cancellationToken = default)
     {
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var responseResult =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
         var getObjectsListResponse = new GetObjectsListResponse(responseResult.StatusCode, responseResult.Content);
         return getObjectsListResponse.ObjectsTuple;
@@ -813,9 +840,10 @@ public partial class MinioClient : IBucketOperations
     private async Task<Tuple<ListVersionsResult, List<Item>>> GetObjectVersionsListAsync(GetObjectListArgs args,
         CancellationToken cancellationToken = default)
     {
-        var requestMessageBuilder = await CreateRequest(args).ConfigureAwait(false);
+        var requestMessageBuilder = await this.CreateRequest(args).ConfigureAwait(false);
         using var responseResult =
-            await this.ExecuteTaskAsync(NoErrorHandlers, requestMessageBuilder, cancellationToken: cancellationToken)
+            await this.ExecuteTaskAsync(ResponseErrorHandlers, requestMessageBuilder,
+                    cancellationToken: cancellationToken)
                 .ConfigureAwait(false);
         var getObjectsVersionsListResponse =
             new GetObjectsVersionsListResponse(responseResult.StatusCode, responseResult.Content);
