+++ Minio/MinioClient.cs
@@ -16,15 +16,9 @@
  */
 
 using System.Net;
-using System.Reflection;
-using System.Runtime.InteropServices;
 using System.Text;
 using CommunityToolkit.HighPerformance;
-using Minio.Credentials;
-using Minio.DataModel;
-using Minio.DataModel.Args;
 using Minio.DataModel.Result;
-using Minio.DataModel.Tracing;
 using Minio.Exceptions;
 using Minio.Handlers;
 using Minio.Helper;
@@ -33,19 +27,9 @@ namespace Minio;
 
 public partial class MinioClient : IMinioClient
 {
-    public MinioConfig Config { get; } = new MinioConfig();
-
     private static readonly char[] separator = { '/' };
 
-    internal readonly IEnumerable<IApiResponseErrorHandler> NoErrorHandlers =
-        Enumerable.Empty<IApiResponseErrorHandler>();
-
     private bool disposedValue;
-
-    private IRequestLogger logger;
-
-    // Enables HTTP tracing if set to true
-    private bool trace;
 
     /// <summary>
     ///     Creates and returns an MinIO Client
@@ -55,12 +39,19 @@ public partial class MinioClient : IMinioClient
     {
     }
 
+    internal IWebProxy Proxy { get; set; }
+
+    public MinioConfig Config { get; } = new();
+
+    public IEnumerable<IApiResponseErrorHandler> ResponseErrorHandlers { get; internal set; } =
+        Enumerable.Empty<IApiResponseErrorHandler>();
+
     /// <summary>
     ///     Default error handling delegate
     /// </summary>
-    public IApiResponseErrorHandler DefaultErrorHandler { get; set; } = new DefaultErrorHandler();
+    public IApiResponseErrorHandler DefaultErrorHandler { get; internal set; } = new DefaultErrorHandler();
 
-    internal IWebProxy Proxy { get; set; }
+    public IRequestLogger Logger { get; internal set; }
 
     /// <summary>
     ///     Runs httpClient's GetAsync method
@@ -83,8 +74,8 @@ public partial class MinioClient : IMinioClient
     /// </summary>
     public void SetTraceOn(IRequestLogger logger = null)
     {
-        this.logger = logger ?? new DefaultRequestLogger();
-        trace = true;
+        Logger = logger ?? new DefaultRequestLogger();
+        Config.TraceHttp = true;
     }
 
     /// <summary>
@@ -92,7 +83,7 @@ public partial class MinioClient : IMinioClient
     /// </summary>
     public void SetTraceOff()
     {
-        trace = false;
+        Config.TraceHttp = false;
     }
 
     public void Dispose()
@@ -100,199 +91,6 @@ public partial class MinioClient : IMinioClient
         // Do not change this code. Put cleanup code in 'Dispose(bool disposing)' method
         Dispose(true);
         GC.SuppressFinalize(this);
-    }
-
-    /// <summary>
-    ///     Resolve region of the bucket.
-    /// </summary>
-    /// <param name="bucketName"></param>
-    /// <returns></returns>
-    private async Task<string> GetRegion(string bucketName)
-    {
-        var rgn = "";
-        // Use user specified region in client constructor if present
-        if (!string.IsNullOrEmpty(Config.Region)) return Config.Region;
-
-        // pick region from endpoint if present
-        if (!string.IsNullOrEmpty(Config.Endpoint))
-            rgn = RegionHelper.GetRegionFromEndpoint(Config.Endpoint);
-
-        // Pick region from location HEAD request
-        if (rgn?.Length == 0)
-            rgn = BucketRegionCache.Instance.Exists(bucketName)
-                ? await BucketRegionCache.Update(this, bucketName).ConfigureAwait(false)
-                : BucketRegionCache.Instance.Region(bucketName);
-
-        // Defaults to us-east-1 if region could not be found
-        return rgn?.Length == 0 ? "us-east-1" : rgn;
-    }
-
-    /// <summary>
-    ///     Null Check for Args object.
-    ///     Expected to be called from CreateRequest
-    /// </summary>
-    /// <param name="args">The child object of Args class</param>
-    private void ArgsCheck(RequestArgs args)
-    {
-        if (args is null)
-            throw new ArgumentNullException(nameof(args),
-                "Args object cannot be null. It needs to be assigned to an instantiated child object of Args.");
-    }
-
-    /// <summary>
-    ///     Constructs a HttpRequestMessageBuilder using bucket/object names from Args.
-    ///     Calls overloaded CreateRequest method.
-    /// </summary>
-    /// <param name="args">The direct descendant of BucketArgs class, args with populated values from Input</param>
-    /// <returns>A HttpRequestMessageBuilder</returns>
-    internal async Task<HttpRequestMessageBuilder> CreateRequest<T>(BucketArgs<T> args) where T : BucketArgs<T>
-    {
-        ArgsCheck(args);
-        var requestMessageBuilder =
-            await CreateRequest(args.RequestMethod, args.BucketName, headerMap: args.Headers,
-                isBucketCreationRequest: args.IsBucketCreationRequest).ConfigureAwait(false);
-        return args.BuildRequest(requestMessageBuilder);
-    }
-
-    /// <summary>
-    ///     Constructs a HttpRequestMessage using bucket/object names from Args.
-    ///     Calls overloaded CreateRequest method.
-    /// </summary>
-    /// <param name="args">The direct descendant of ObjectArgs class, args with populated values from Input</param>
-    /// <returns>A HttpRequestMessage</returns>
-    internal async Task<HttpRequestMessageBuilder> CreateRequest<T>(ObjectArgs<T> args) where T : ObjectArgs<T>
-    {
-        ArgsCheck(args);
-
-        var contentType = "application/octet-stream";
-        _ = args.Headers?.TryGetValue("Content-Type", out contentType);
-        var requestMessageBuilder =
-            await CreateRequest(args.RequestMethod,
-                args.BucketName,
-                args.ObjectName,
-                args.Headers,
-                contentType,
-                args.RequestBody).ConfigureAwait(false);
-        return args.BuildRequest(requestMessageBuilder);
-    }
-
-    /// <summary>
-    ///     Constructs an HttpRequestMessage builder. For AWS, this function
-    ///     has the side-effect of overriding the baseUrl in the HttpClient
-    ///     with region specific host path or virtual style path.
-    /// </summary>
-    /// <param name="method">HTTP method</param>
-    /// <param name="bucketName">Bucket Name</param>
-    /// <param name="objectName">Object Name</param>
-    /// <param name="headerMap">headerMap</param>
-    /// <param name="contentType">Content Type</param>
-    /// <param name="body">request body</param>
-    /// <param name="resourcePath">query string</param>
-    /// <param name="isBucketCreationRequest">boolean to define bucket creation</param>
-    /// <returns>A HttpRequestMessage builder</returns>
-    /// <exception cref="BucketNotFoundException">When bucketName is invalid</exception>
-    internal async Task<HttpRequestMessageBuilder> CreateRequest(
-        HttpMethod method,
-        string bucketName = null,
-        string objectName = null,
-        IDictionary<string, string> headerMap = null,
-        string contentType = "application/octet-stream",
-        ReadOnlyMemory<byte> body = default,
-        string resourcePath = null,
-        bool isBucketCreationRequest = false)
-    {
-        var region = string.Empty;
-        if (bucketName is not null)
-        {
-            Utils.ValidateBucketName(bucketName);
-            // Fetch correct region for bucket if this is not a bucket creation
-            if (!isBucketCreationRequest)
-                region = await GetRegion(bucketName).ConfigureAwait(false);
-        }
-
-        if (objectName is not null) Utils.ValidateObjectName(objectName);
-
-        if (Config.Provider is not null)
-        {
-            var isAWSEnvProvider = Config.Provider is AWSEnvironmentProvider ||
-                                   (Config.Provider is ChainedProvider ch &&
-                                    ch.CurrentProvider is AWSEnvironmentProvider);
-
-            var isIAMAWSProvider = Config.Provider is IAMAWSProvider ||
-                                   (Config.Provider is ChainedProvider chained &&
-                                    chained.CurrentProvider is AWSEnvironmentProvider);
-
-            AccessCredentials creds;
-            if (isAWSEnvProvider)
-            {
-                var aWSEnvProvider = (AWSEnvironmentProvider)Config.Provider;
-                creds = await aWSEnvProvider.GetCredentialsAsync().ConfigureAwait(false);
-            }
-            else if (isIAMAWSProvider)
-            {
-                var iamAWSProvider = (IAMAWSProvider)Config.Provider;
-                creds = iamAWSProvider.Credentials;
-            }
-            else
-            {
-                creds = await Config.Provider.GetCredentialsAsync().ConfigureAwait(false);
-            }
-
-            if (creds is not null)
-            {
-                Config.AccessKey = creds.AccessKey;
-                Config.SecretKey = creds.SecretKey;
-            }
-        }
-
-        // This section reconstructs the url with scheme followed by location specific endpoint (s3.region.amazonaws.com)
-        // or Virtual Host styled endpoint (bucketname.s3.region.amazonaws.com) for Amazon requests.
-        var resource = string.Empty;
-        var usePathStyle = false;
-
-        if (!string.IsNullOrEmpty(bucketName) && S3utils.IsAmazonEndPoint(Config.BaseUrl))
-        {
-            if (method == HttpMethod.Put && objectName is null && resourcePath is null)
-                // use path style for make bucket to workaround "AuthorizationHeaderMalformed" error from s3.amazonaws.com
-                usePathStyle = true;
-            else if (resourcePath?.Contains("location", StringComparison.OrdinalIgnoreCase) == true)
-                // use path style for location query
-                usePathStyle = true;
-            else if (bucketName.Contains('.', StringComparison.Ordinal) && Config.Secure)
-                // use path style where '.' in bucketName causes SSL certificate validation error
-                usePathStyle = true;
-
-            if (usePathStyle) resource += Utils.UrlEncode(bucketName) + "/";
-        }
-
-        // Set Target URL
-        var requestUrl = RequestUtil.MakeTargetURL(Config.BaseUrl, Config.Secure, bucketName, region, usePathStyle);
-
-        if (objectName is not null) resource += Utils.EncodePath(objectName);
-
-        // Append query string passed in
-        if (resourcePath is not null) resource += resourcePath;
-
-        HttpRequestMessageBuilder messageBuilder;
-        if (!string.IsNullOrEmpty(resource))
-            messageBuilder = new HttpRequestMessageBuilder(method, requestUrl, resource);
-        else
-            messageBuilder = new HttpRequestMessageBuilder(method, requestUrl);
-        if (!body.IsEmpty)
-        {
-            messageBuilder.SetBody(body);
-            messageBuilder.AddOrUpdateHeaderParameter("Content-Type", contentType);
-        }
-
-        if (headerMap is not null)
-        {
-            if (headerMap.TryGetValue(messageBuilder.ContentTypeKey, out var value) && !string.IsNullOrEmpty(value))
-                headerMap[messageBuilder.ContentTypeKey] = contentType;
-
-            foreach (var entry in headerMap) messageBuilder.AddOrUpdateHeaderParameter(entry.Key, entry.Value);
-        }
-
-        return messageBuilder;
     }
 
     /// <summary>
@@ -487,71 +285,6 @@ public partial class MinioClient : IMinioClient
                 nameof(response));
 
         throw new UnexpectedMinioException(errResponse.Message) { Response = errResponse, XmlError = response.Content };
-    }
-
-    /// <summary>
-    ///     Delegate errors to handlers
-    /// </summary>
-    /// <param name="response"></param>
-    /// <param name="handlers"></param>
-    /// <param name="startTime"></param>
-    private void HandleIfErrorResponse(ResponseResult response, IEnumerable<IApiResponseErrorHandler> handlers,
-        DateTime startTime)
-    {
-        // Logs Response if HTTP tracing is enabled
-        if (trace)
-        {
-            var now = DateTime.Now;
-            LogRequest(response.Request, response, (now - startTime).TotalMilliseconds);
-        }
-
-        if (handlers is null) throw new ArgumentNullException(nameof(handlers));
-
-        // Run through handlers passed to take up error handling
-        foreach (var handler in handlers) handler.Handle(response);
-
-        // Fall back default error handler
-        DefaultErrorHandler.Handle(response);
-    }
-
-    /// <summary>
-    ///     Logs the request sent to server and corresponding response
-    /// </summary>
-    /// <param name="request"></param>
-    /// <param name="response"></param>
-    /// <param name="durationMs"></param>
-    private void LogRequest(HttpRequestMessage request, ResponseResult response, double durationMs)
-    {
-        var requestToLog = new RequestToLog
-        {
-            Resource = request.RequestUri.PathAndQuery,
-            // Parameters are custom anonymous objects in order to have the parameter type as a nice string
-            // otherwise it will just show the enum value
-            Parameters = request.Headers.Select(parameter => new RequestParameter
-            {
-                Name = parameter.Key,
-                Value = parameter.Value,
-                Type = typeof(KeyValuePair<string, IEnumerable<string>>).ToString()
-            }),
-            // ToString() here to have the method as a nice string otherwise it will just show the enum value
-            Method = request.Method.ToString(),
-            // This will generate the actual Uri used in the request
-            Uri = request.RequestUri
-        };
-
-        var responseToLog = new ResponseToLog
-        {
-            StatusCode = response.StatusCode,
-            Content = response.Content,
-            Headers = response.Headers.ToDictionary(o => o.Key, o => string.Join(Environment.NewLine, o.Value),
-                StringComparer.Ordinal),
-            // The Uri that actually responded (could be different from the requestUri if a redirection occurred)
-            ResponseUri = response.Request.RequestUri,
-            ErrorMessage = response.ErrorMessage,
-            DurationMs = durationMs
-        };
-
-        logger.LogRequest(requestToLog, responseToLog, durationMs);
     }
 
     protected virtual void Dispose(bool disposing)
